sergeai-backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation
pydantic[email]==2.5.0
email-validator==2.1.0

# AI Integration (Optional)
openai==1.3.5

# For development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3